This is yet another iteration on top of https://github.com/subspace/subspace/pull/1683 that primarily benefits larger k, on k22 I've got these results (~100ms less than https://github.com/subspace/subspace/pull/1683 when using all cores, single-core didn't change meaningfully):
```
chia/table/parallel     time:   [605.64 ms 619.75 ms 633.06 ms]
```

Yeah, I'd be able to farm with k22 on my machine now :sunglasses: 

On 4 performance cores we're looking at this:
```
chia/table/parallel     time:   [1.4368 s 1.4439 s 1.4510 s]
```

So even with slower processor it should still be well within 6 seconds window.

---

I have tried various additional SIMD and parallelization technics, but looks like I'm hitting hardware limitations. Things either stay the same (meaning compiler/CPU is doing as well of a job with simpler code) or worse. This is as far as I can get it without resorting to lower-level architecture-specific intrinsics, which is something I'm not good at.

In particular `perf` says application is 30%+ memory bound and has 50%+ cache misses on my CPU, branch misprediction and bad speculation are also at ~10% level:
```
cpu_core/topdown-retiring/u      # 35,1% Retiring           (46,90%)
cpu_core/topdown-bad-spec/u      #  8,6% Bad Speculation    (46,90%)
cpu_core/topdown-fe-bound/u      #  9,8% Frontend Bound     (46,90%)
cpu_core/topdown-be-bound/u      # 46,5% Backend Bound      (46,90%)
cpu_core/topdown-heavy-ops/u     #  1,4% Heavy Operations  #     33,7% Light Operations (46,90%)
cpu_core/topdown-br-mispredict/u #  8,2% Branch Mispredict #      0,4% Machine Clears   (46,90%)
cpu_core/topdown-fetch-lat/u     #  7,1% Fetch Latency     #      2,7% Fetch Bandwidth  (46,90%)
cpu_core/topdown-mem-bound/u     # 34,3% Memory Bound      #     12,2% Core Bound       (46,90%)
cpu_core/cache-misses:u/                                    (48,04%)
cpu_atom/cache-misses:u/                                    (52,11%)
```

I have one more thing to do that is not algorithm improvement, but rather API change for better cache reuse, but that will be a separate PR.

### Code contributor checklist:
* [x] I have read, understood and followed [contributing guide](https://github.com/subspace/subspace/blob/main/CONTRIBUTING.md)